+++ absl/flags/_flagvalues.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 """Defines the FlagValues class - registry of 'Flag' objects.
 
 Do NOT import this module directly. Import the flags package and use the
@@ -33,6 +32,14 @@ from absl.flags import _exceptions
 from absl.flags import _flag
 from absl.flags import _helpers
 import six
+
+# pylint: disable=unused-import
+try:
+  import typing
+  from typing import Text, Optional
+except ImportError:
+  typing = None
+# pylint: enable=unused-import
 
 # Add flagvalues module to disclaimed module ids.
 _helpers.disclaim_module_ids.add(id(sys.modules[__name__]))
@@ -221,8 +228,7 @@ class FlagValues(object):
       return True
     # Check whether flag_obj is registered under its short name.
     short_name = flag_obj.short_name
-    if (short_name is not None and
-        flag_dict.get(short_name, None) == flag_obj):
+    if (short_name is not None and flag_dict.get(short_name, None) == flag_obj):
       return True
     return False
 
@@ -246,7 +252,7 @@ class FlagValues(object):
         while flag_obj in flags_in_module:
           flags_in_module.remove(flag_obj)
 
-  def _get_flags_defined_by_module(self, module):
+  def get_flags_for_module(self, module):
     """Returns the list of flags defined by a module.
 
     Args:
@@ -259,6 +265,8 @@ class FlagValues(object):
     """
     if not isinstance(module, str):
       module = module.__name__
+    if module == '__main__':
+      module = sys.argv[0]
 
     return list(self.flags_by_module_dict().get(module, []))
 
@@ -275,11 +283,13 @@ class FlagValues(object):
     """
     if not isinstance(module, str):
       module = module.__name__
+    if module == '__main__':
+      module = sys.argv[0]
 
     # Any flag is a key flag for the module that defined it.  NOTE:
     # key_flags is a fresh list: we can update it without affecting the
     # internals of this FlagValues object.
-    key_flags = self._get_flags_defined_by_module(module)
+    key_flags = self.get_flags_for_module(module)
 
     # Take into account flags explicitly declared as key for a module.
     for flag in self.key_flags_by_module_dict().get(module, []):
@@ -292,8 +302,7 @@ class FlagValues(object):
 
     Args:
       flagname: str, name of the flag to lookup.
-      default: Value to return if flagname is not defined. Defaults
-          to None.
+      default: Value to return if flagname is not defined. Defaults to None.
 
     Returns:
       The name of the module which registered the flag with this name.
@@ -318,8 +327,7 @@ class FlagValues(object):
 
     Args:
       flagname: str, name of the flag to lookup.
-      default: Value to return if flagname is not defined. Defaults
-          to None.
+      default: Value to return if flagname is not defined. Defaults to None.
 
     Returns:
       The ID of the module which registered the flag with this name.
@@ -343,8 +351,8 @@ class FlagValues(object):
     """Allow set default values for undefined flags.
 
     Args:
-      setter: Method(name, value) to call to __setattr__ an unknown flag.
-        Must raise NameError or ValueError for invalid name/value.
+      setter: Method(name, value) to call to __setattr__ an unknown flag. Must
+        raise NameError or ValueError for invalid name/value.
     """
     self.__dict__['__set_unknown'] = setter
 
@@ -369,7 +377,7 @@ class FlagValues(object):
         return value
       except (TypeError, ValueError):  # Flag value is not valid.
         raise _exceptions.IllegalFlagValueError(
-            '"{1}" is not valid for --{0}' .format(name, value))
+            '"{1}" is not valid for --{0}'.format(name, value))
       except NameError:  # Flag name is not valid.
         pass
     raise _exceptions.UnrecognizedFlagError(name, value)
@@ -398,7 +406,7 @@ class FlagValues(object):
 
     Args:
       flag_values: FlagValues, the FlagValues instance containing flags to
-          remove.
+        remove.
     """
     for flag_name in flag_values:
       self.__delattr__(flag_name)
@@ -440,8 +448,7 @@ class FlagValues(object):
         flags_to_cleanup.add(fl[short_name])
       fl[short_name] = flag
     if (name not in fl  # new flag
-        or fl[name].using_default_value
-        or not flag.using_default_value):
+        or fl[name].using_default_value or not flag.using_default_value):
       if name in fl and fl[name] != flag:
         flags_to_cleanup.add(fl[name])
       fl[name] = flag
@@ -477,8 +484,8 @@ class FlagValues(object):
     if self.__dict__['__flags_parsed'] or fl[name].present:
       return fl[name].value
     else:
-      error_message = (
-          'Trying to access flag --%s before flags were parsed.' % name)
+      error_message = ('Trying to access flag --%s before flags were parsed.' %
+                       name)
       if six.PY2:
         # In Python 2, hasattr returns False if getattr raises any exception.
         # That means if someone calls hasattr(FLAGS, 'flag'), it returns False
@@ -502,11 +509,17 @@ class FlagValues(object):
     fl[name].using_default_value = False
     return value
 
-  def _assert_all_validators(self):
+  def validate_all_flags(self):
+    """Verifies whether all flags pass validation.
+
+    Raises:
+      AttributeError: Raised if validators work with a non-existing flag.
+      IllegalFlagValueError: Raised if validation fails for at least one
+          validator.
+    """
     all_validators = set()
     for flag in six.itervalues(self._flags()):
-      for validator in flag.validators:
-        all_validators.add(validator)
+      all_validators.update(flag.validators)
     self._assert_validators(all_validators)
 
   def _assert_validators(self, validators):
@@ -515,8 +528,8 @@ class FlagValues(object):
     It asserts validators in the order they were created.
 
     Args:
-      validators: Iterable(validators.Validator), validators to be
-          verified.
+      validators: Iterable(validators.Validator), validators to be verified.
+
     Raises:
       AttributeError: Raised if validators work with a non-existing flag.
       IllegalFlagValueError: Raised if validation fails for at least one
@@ -600,7 +613,7 @@ class FlagValues(object):
     Args:
        argv: a tuple/list of strings.
        known_only: bool, if True, parse and remove known flags; return the rest
-           untouched. Unknown flags specified by --undefok are not returned.
+         untouched. Unknown flags specified by --undefok are not returned.
 
     Returns:
        The list of arguments not parsed as options, including argv[0].
@@ -633,7 +646,7 @@ class FlagValues(object):
           name, value, suggestions=suggestions)
 
     self.mark_as_parsed()
-    self._assert_all_validators()
+    self.validate_all_flags()
     return [program_name] + unparsed_args
 
   def __getstate__(self):
@@ -656,7 +669,7 @@ class FlagValues(object):
 
     Args:
       is_retired_flag_func: Callable(str) -> (bool, bool), a function takes flag
-          name as parameter, returns a tuple (is_retired, type_is_bool).
+        name as parameter, returns a tuple (is_retired, type_is_bool).
     """
     self.__dict__['__is_retired_flag_func'] = is_retired_flag_func
 
@@ -670,7 +683,7 @@ class FlagValues(object):
     Args:
       args: [str], a list of strings with the arguments to parse.
       known_only: bool, if True, parse and remove known flags; return the rest
-           untouched. Unknown flags specified by --undefok are not returned.
+        untouched. Unknown flags specified by --undefok are not returned.
 
     Returns:
       A tuple with the following:
@@ -766,8 +779,9 @@ class FlagValues(object):
             # This happens when a non-bool retired flag is specified
             # in format of "--flag value".
             get_value()
-          logging.error('Flag "%s" is retired and should no longer '
-                        'be specified. See go/totw/90.', name)
+          logging.error(
+              'Flag "%s" is retired and should no longer '
+              'be specified. See go/totw/90.', name)
           continue
 
       if flag:
@@ -851,8 +865,9 @@ class FlagValues(object):
       # Just print one long list of flags.
       values = six.itervalues(self._flags())
       if include_special_flags:
-        values = itertools.chain(
-            values, six.itervalues(_helpers.SPECIAL_FLAGS._flags()))  # pylint: disable=protected-access
+        values = itertools.chain(values,
+                                 six.itervalues(
+                                     _helpers.SPECIAL_FLAGS._flags()))  # pylint: disable=protected-access
       self._render_flag_list(values, output_lines, prefix)
       return '\n'.join(output_lines)
 
@@ -887,7 +902,7 @@ class FlagValues(object):
 
   def _render_our_module_flags(self, module, output_lines, prefix=''):
     """Returns a help string for a given module."""
-    flags = self._get_flags_defined_by_module(module)
+    flags = self.get_flags_for_module(module)
     if flags:
       self._render_module_flags(module, flags, output_lines, prefix)
 
@@ -896,8 +911,8 @@ class FlagValues(object):
 
     Args:
       module: module|str, the module to render key flags for.
-      output_lines: [str], a list of strings.  The generated help message
-          lines will be appended to this list.
+      output_lines: [str], a list of strings.  The generated help message lines
+        will be appended to this list.
       prefix: str, a string that is prepended to each generated help line.
     """
     key_flags = self.get_key_flags_for_module(module)
@@ -939,10 +954,12 @@ class FlagValues(object):
         # a different flag is using this name now
         continue
       # only print help once
-      if flag in flagset: continue
+      if flag in flagset:
+        continue
       flagset[flag] = 1
       flaghelp = ''
-      if flag.short_name: flaghelp += '-%s,' % flag.short_name
+      if flag.short_name:
+        flaghelp += '-%s,' % flag.short_name
       if flag.boolean:
         flaghelp += '--[no]%s:' % flag.name
       else:
@@ -951,15 +968,15 @@ class FlagValues(object):
       if flag.help:
         flaghelp += flag.help
       flaghelp = _helpers.text_wrap(
-          flaghelp, indent=prefix+'  ', firstline_indent=prefix)
+          flaghelp, indent=prefix + '  ', firstline_indent=prefix)
       if flag.default_as_str:
         flaghelp += '\n'
         flaghelp += _helpers.text_wrap(
-            '(default: %s)' % flag.default_as_str, indent=prefix+'  ')
+            '(default: %s)' % flag.default_as_str, indent=prefix + '  ')
       if flag.parser.syntactic_help:
         flaghelp += '\n'
         flaghelp += _helpers.text_wrap(
-            '(%s)' % flag.parser.syntactic_help, indent=prefix+'  ')
+            '(%s)' % flag.parser.syntactic_help, indent=prefix + '  ')
       output_lines.append(flaghelp)
 
   def get_flag_value(self, name, default):  # pylint: disable=invalid-name
@@ -1015,8 +1032,7 @@ class FlagValues(object):
     elif flagfile_str.startswith('-flagfile='):
       return os.path.expanduser((flagfile_str[(len('-flagfile=')):]).strip())
     else:
-      raise _exceptions.Error(
-          'Hit illegal --flagfile type: %s' % flagfile_str)
+      raise _exceptions.Error('Hit illegal --flagfile type: %s' % flagfile_str)
 
   def _get_flag_file_lines(self, filename, parsed_file_stack=None):
     """Returns the useful (!=comments, etc) lines from a file with flags.
@@ -1090,11 +1106,11 @@ class FlagValues(object):
 
     Args:
       argv: [str], a list of strings, usually sys.argv[1:], which may contain
-          one or more flagfile directives of the form --flagfile="./filename".
-          Note that the name of the program (sys.argv[0]) should be omitted.
+        one or more flagfile directives of the form --flagfile="./filename".
+        Note that the name of the program (sys.argv[0]) should be omitted.
       force_gnu: bool, if False, --flagfile parsing obeys the
-          FLAGS.is_gnu_getopt() value. If True, ignore the value and always
-          follow gnu_getopt semantics.
+        FLAGS.is_gnu_getopt() value. If True, ignore the value and always follow
+        gnu_getopt semantics.
 
     Returns:
       A new list which has the original list combined with what we read
@@ -1151,8 +1167,8 @@ class FlagValues(object):
           if not force_gnu and not self.__dict__['__use_gnu_getopt']:
             break
         else:
-          if ('=' not in current_arg and
-              rest_of_args and not rest_of_args[0].startswith('-')):
+          if ('=' not in current_arg and rest_of_args and
+              not rest_of_args[0].startswith('-')):
             # If this is an occurrence of a legitimate --x y, skip the value
             # so that it won't be mistaken for a standalone arg.
             fl = self._flags()
@@ -1220,16 +1236,17 @@ class FlagValues(object):
     all_flag = doc.createElement('AllFlags')
     doc.appendChild(all_flag)
 
-    all_flag.appendChild(_helpers.create_xml_dom_element(
-        doc, 'program', os.path.basename(sys.argv[0])))
+    all_flag.appendChild(
+        _helpers.create_xml_dom_element(doc, 'program',
+                                        os.path.basename(sys.argv[0])))
 
     usage_doc = sys.modules['__main__'].__doc__
     if not usage_doc:
       usage_doc = '\nUSAGE: %s [flags]\n' % sys.argv[0]
     else:
       usage_doc = usage_doc.replace('%s', sys.argv[0])
-    all_flag.appendChild(_helpers.create_xml_dom_element(
-        doc, 'usage', usage_doc))
+    all_flag.appendChild(
+        _helpers.create_xml_dom_element(doc, 'usage', usage_doc))
 
     # Get list of key flags for the main module.
     key_flags = self.get_key_flags_for_module(sys.argv[0])
@@ -1243,8 +1260,11 @@ class FlagValues(object):
       flag_list.sort()
       for unused_flag_name, flag in flag_list:
         is_key = flag in key_flags
-        all_flag.appendChild(flag._create_xml_dom_element(  # pylint: disable=protected-access
-            doc, module_name, is_key=is_key))
+        all_flag.appendChild(
+            flag._create_xml_dom_element(  # pylint: disable=protected-access
+                doc,
+                module_name,
+                is_key=is_key))
 
     outfile = outfile or sys.stdout
     if six.PY2:
@@ -1271,3 +1291,84 @@ class FlagValues(object):
 
 
 FLAGS = FlagValues()
+
+if typing:
+  _T = typing.TypeVar('_T')
+  _Base = typing.Generic[_T]
+else:
+  _Base = object
+
+
+class FlagHolder(_Base):
+  """Holds a defined flag.
+
+  This facilitates a cleaner api around global state. Instead of
+
+  ```
+  flags.DEFINE_integer('foo', ...)
+  flags.DEFINE_integer('bar', ...)
+  ...
+  def method():
+    # prints parsed value of 'bar' flag
+    print(flags.FLAGS.foo)
+    # runtime error due to typo or possibly bad coding style.
+    print(flags.FLAGS.baz)
+  ```
+
+  it encourages code like
+
+  ```
+  FOO_FLAG = flags.DEFINE_integer('foo', ...)
+  BAR_FLAG = flags.DEFINE_integer('bar', ...)
+  ...
+  def method():
+    print(FOO_FLAG.value)
+    print(BAR_FLAG.value)
+  ```
+
+  since the name of the flag appears only once in the source code.
+  """
+
+  def __init__(self, flag_values, flag, ensure_non_none_value=False):
+    """Constructs a FlagHolder instance providing typesafe access to flag.
+
+    Args:
+      flag_values: The container the flag is registered to.
+      flag: The flag object for this flag.
+      ensure_non_none_value: Is the value of the flag allowed to be None.
+    """
+    self._flagvalues = flag_values
+    # We take the entire flag object, but only keep the name. Why?
+    # - We want FlagHolder[T] to be generic container
+    # - flag_values contains all flags, so has no reference to T.
+    # - typecheckers don't like to see a generic class where none of the ctor
+    #   arguments refer to the generic type.
+    self._name = flag.name
+    # We intentionally do NOT check if the default value is None.
+    # This allows future use of this for "required flags with None default"
+    self._ensure_non_none_value = ensure_non_none_value
+
+  @property
+  def name(self):
+    return self._name
+
+  @property
+  def value(self):
+    """Returns the value of the flag.
+
+    If _ensure_non_none_value is True, then return value is not None.
+
+    Raises:
+      UnparsedFlagAccessError: if flag parsing has not finished.
+      IllegalFlagValueError: if value is None unexpectedly.
+    """
+    val = getattr(self._flagvalues, self._name)
+    if self._ensure_non_none_value and val is None:
+      raise _exceptions.IllegalFlagValueError(
+          'Unexpected None value for flag %s' % self._name)
+    return val
+
+  @property
+  def default(self):
+    """Returns the default value of the flag."""
+    return self._flagvalues[self._name].default